% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_resp_ind.r
\name{dose_resp_ind}
\alias{dose_resp_ind}
\title{Extract the information from the simulation data frame to analyse the dose response effects}
\usage{
dose_resp_ind(allsim, dataset, dr = seq(0, 1, 0.1))
}
\arguments{
\item{allsim}{dataset with all simulations values}

\item{dataset}{dataset with all variables}

\item{dr}{a vector with dose response values}
}
\value{
a data frame with dose response values
}
\description{
Extract the information from the simulation data frame to analyse the dose response effects
}
