% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_data.r
\docType{data}
\name{expose_data}
\alias{expose_data}
\title{Multiple quemicals and several epidemiological outcomes}
\format{An object of class \code{'data frame'}.}
\usage{
data(expose_data)
}
\description{
Data from a epidemiological study on multiple-pollutants
Dataset with 300 records each one representing an individual and 11 variables:
5 quemicals, gender, 4 measured outcomes and Id.
}
\examples{
data(expose_data)
head(expose_data)
}
\keyword{datasets}
