% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.r
\docType{data}
\name{gen}
\alias{gen}
\title{One of the intermediate data frame}
\format{An object of class \code{'data frame'}.}
\usage{
data(gen)
}
\description{
Data from a epidemiological study on multiple-pollutants
Dataset with 300 records each one representing an individual and 11 variables:
5 quemicals, gender, 4 measured outcomes and Id.
}
\examples{
data(gen)
head(gen)
}
\keyword{datasets}
