% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_ace_ind.r
\name{naive_ace_ind}
\alias{naive_ace_ind}
\title{Extract the information from the simulation data frame to analyse the interaction effects}
\usage{
naive_ace_ind(allsim, dataset, ic_dis = "IC")
}
\arguments{
\item{allsim}{dataset with all simulations values}

\item{dataset}{dataset with all variables}

\item{ic_dis}{choose between ic (interval confidences) and dis (distribution)}
}
\value{
a data frame with naive ace and confident intervals
}
\description{
Extract the information from the simulation data frame to analyse the interaction effects
}
