% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.r
\name{run_simulations}
\alias{run_simulations}
\title{This function run all the simulations based on the structure of the basic data frame.}
\usage{
run_simulations(dataset, exposures, confounders, libraries, outcomes,
  num.sim = 50, delta = c(0, 1), dr, newdata, show_times = FALSE,
  show_num_sim = TRUE, save_time = FALSE, verbose = FALSE,
  family = "gaussian", method = "method.NNLS")
}
\arguments{
\item{dataset}{dataset with all variables}

\item{exposures}{a vector with exposures}

\item{confounders}{a vector with confounders}

\item{libraries}{a vector of libraries to use}

\item{outcomes}{outcome´s name}

\item{num.sim}{number of simulations}

\item{delta}{a vector with two values}

\item{dr}{a vector with dose response values}

\item{newdata}{the dataframe with new values}

\item{show_times}{boolean parameter to see the time}

\item{show_num_sim}{boolean parameter to see the iteration of simulations}

\item{save_time}{boolean parameter to save the time in the result list}

\item{verbose}{boolean parameter to see the verbose of superlearner}

\item{family}{a character parameter to describe the family of the model}

\item{method}{a character parameter to choose the method in the superlearner}
}
\value{
a list with 4 objects: a data frame with all simulations, risk and coefficients of the crossvalidation and the time of the proces.
}
\description{
This function run all the simulations based on the structure of the basic data frame.
}
\details{
libraries could be SL if we don't select nothing or 'SL.glm', 'SL.glm.interaction','SL.glmnet', 'SL.gam','SL.xgboost','SL.polymars','SL.randomForest'
}
