% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_resp.r
\name{dose_resp}
\alias{dose_resp}
\title{Extract the information from the simulation data frame to analyse the dose response effects}
\usage{
dose_resp(allsim, dataset, exposures, dr, ic_dis = "IC", st)
}
\arguments{
\item{allsim}{dataset with all simulations values}

\item{dataset}{dataset with all variables}

\item{exposures}{a vector with exposures}

\item{dr}{a vector with dose response values}

\item{ic_dis}{choose between ic (interval confidences) and dis (distribution)}

\item{st}{summary table from general function}
}
\value{
a data frame with dose response values
}
\description{
Extract the information from the simulation data frame to analyse the dose response effects
}
