% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.r
\name{general_function}
\alias{general_function}
\title{General function}
\usage{
general_function(dataset, exposures, confounders, outcomes, delta, dr)
}
\arguments{
\item{dataset}{dataset with all variables}

\item{exposures}{a vector with exposures}

\item{confounders}{a vector with confounders}

\item{outcomes}{outcome´s name}

\item{delta}{a vector with two values}

\item{dr}{a vector with dose response values}
}
\value{
a list with 2 objects. One is the dataframe with all the values and the other is a summary of the groups and the corresponding rows in the first dataframe.
}
\description{
General function
}
\details{
libraries could be SL if we don't select nothing or 'SL.glm', 'SL.glm.interaction','SL.glmnet', 'SL.gam','SL.xgboost','SL.polymars','SL.randomForest'
}
