% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.r
\name{ice}
\alias{ice}
\title{Extract the information from the simulation data frame to analyse the individual causal effects}
\usage{
ice(allsim, dataset, dr = seq(0, 1, 0.1), squem, remove_extrem = FALSE)
}
\arguments{
\item{allsim}{dataset with all simulations values}

\item{dataset}{dataset with all variables}

\item{dr}{a vector with dose response values}

\item{squem}{squeme of the values of the prediction values}

\item{remove_extrem}{boolean parameter to remove the extrem values}
}
\value{
a data frame with interactions
}
\description{
Extract the information from the simulation data frame to analyse the individual causal effects
}
