% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.r
\name{interact}
\alias{interact}
\title{Extract the information from the simulation data frame to analyse the interaction effects}
\usage{
interact(allsim, dataset, exposures, confounders, squem)
}
\arguments{
\item{allsim}{dataset with all simulations values}

\item{dataset}{dataset with all variables}

\item{exposures}{a vector with exposures}

\item{confounders}{a vector with confounders}

\item{squem}{squeme of the values of the prediction values}
}
\value{
data frame with interaction values
}
\description{
Extract the information from the simulation data frame to analyse the interaction effects
}
