% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_ace.r
\name{naive_ace}
\alias{naive_ace}
\title{Naive Average causal effect all together}
\usage{
naive_ace(allsim, dataset, exposures, delta = c(0, 1), ic_dis = "IC", st)
}
\arguments{
\item{allsim}{dataset with all simulations values}

\item{dataset}{dataset with all variables}

\item{exposures}{a vector with exposures}

\item{delta}{a vector with two values}

\item{ic_dis}{choose between ic (interval confidences) and dis (distribution)}

\item{st}{summary table from general function}
}
\value{
a data frame with naive ace and confident intervals
}
\description{
Naive Average causal effect all together
}
